Major commit

+++ config/config.go
@@ -6,11 +6,12 @@ import (
 )
 
 var BROADCAST_IP = net.ParseIP("127.255.255.255")
+
 const PORT = 41784
 
 const NUM_ELEVATORS = 3
 const NUM_FLOORS = 4
 
-func LocalIp(id byte) net.IP {
+func LocalIp(id int) net.IP {
 	return net.ParseIP(fmt.Sprintf("127.0.0.%v", id))
 }
